// app/_types/types.ts
export type State = {
  label: string;
  reportFormat: string;
  email: string; // email address
  fileName: string;
  filePath: string;
  description: string;
  hasChanges: boolean;
  removeFile: boolean;
};

export type Action =
  | { type: "ready"; payload: FlowNode | undefined }
  | { type: "setLabel"; payload: string }
  | { type: "setDescription"; payload: string }
  | { type: "setReportFormat"; payload: string }
  | { type: "setEmail"; payload: string }
  | { type: "setFile"; payload: { fileName: string; filePath: string } }
  | { type: "removeFile" }
  | { type: "reset"; payload: FlowNode | undefined };